docs: Add docs site
When I upgraded Storybook to v7, Docs Addon stopped reading JSDoc
comments. The feature, along with UX for the Docs Page (as an end-user)
have not been great from the start. So I decided to write a document
site from scratch, as the complexity of this project is not high.

Upon writing the docs, I realized I no longer need Storybook anymore
thanks to simple playground section.

* Figspec elements are Custom Element: no need for style isolation
  and it just works with `<iframe>`.
* Minimum HTML+JS+class-less CSS is very lightweight, simple and
  robust. Goodbye monstrously huge dependencies.
* Because how Custom Elements work, HMR is not a thing. Reload is always
  required and this simple docs site is way faster at the page startup
  speed.
* I'm dubious about web-component-analyzer's future and current quality.
  Years passed but it's still "for experimental and demo purposes".
  In the end, burden of duplicated documentation because of Mixins and
  lack of stability outweighed the merit.

Simple `<iframe>` and HTML/JS per stories solved my problem.
It does not allow me to change parameter (Args, in Storybook) but I've
never used the feature in this project. For Viewport Addon
functionality, just open the frame in new tab then use browsers'
built-in device emulation. Many devs knowing how to use Storybook
but it can't beat the benefit comes from the reduction of the number
of project's dependencies by removing SB packages.

So I copied stories to "Examples". (technically it's test cases?)
Removal of Storybook and its files are not included in this commit in
order to keep commits' atomicity.