feat: v2 rewrite
The goals of this rewrite are:

* Improve performance, especially on many nodes.
* Get rid of Lit.
* Make zero-dependencies.
* Make UI useable and ok-ish.

The current quality is not personally satisfiable. Not able to controll
with touch input, still no feedback for copy actions, many missed
opportunity for CSS codegen, etc...
However, I believe this state achieves the above goals, so decided to
create a release instead of perfecting.