chore: Support only ESM-ES2019
The reason `target` set to `ES2015` was I thought "if it's 5yrs old,
it's safe". For the same reason, I tried to set to ES2018 but decided to
increase because:

* Even the Safari 12 mostly supports ES2019.
* Library code uses ES2019 API (`Object.fromEntries` and
  `Array.prototype.flat`).
* There was no differences in built files, between `target=es2018` and
  `target=es2019`.

The reason for dropping CJS is it's obsolescence. It was still used
for front-end code in 3 years ago (when I initially wrote this lib).
However, given the support status of bundlers and Node.js, I believe
dropping CJS output provides better user experience thanks to halved
package size and more maintainability.